+++ lustre/mdt/mdt_handler.c
@@ -451,7 +451,7 @@ void mdt_pack_attr2body(struct mdt_thread_info *info, struct mdt_body *b,
 
         if (b->valid & OBD_MD_FLSIZE)
                 CDEBUG(D_VFSTRACE, DFID": returning size %llu\n",
-                       PFID(fid), b->size);
+                       PFID(fid), (unsigned long long)b->size);
 }
 
 static inline int mdt_body_has_lov(const struct lu_attr *la,
